src/requirements.txt
asgiref==3.8.1
chardet==5.2.0
dj-database-url==2.2.0
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
Faker==26.1.0
gunicorn==22.0.0
inflection==0.5.1
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
reportlab==4.2.2
setuptools==70.3.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
